Move logic around, so we don't need to import the plugin.
TODO: Still need to expose these to settings
